refactor: modularize navbar sections
